changing percentage to raw number
