+++ pom.xml
@@ -7,7 +7,7 @@
     <groupId>com.flipkart.ranger</groupId>
     <artifactId>ranger</artifactId>
     <packaging>jar</packaging>
-    <version>0.2-BETA-SNAPSHOT</version>
+    <version>0.3-BETA-SNAPSHOT</version>
 
     <distributionManagement>
         <repository>
